resources/health-analyzer/requirements.txt
neurokit2>=0.2.7
py-ecg-detectors>=1.3.4
pandas>=2.0
numpy<2
scipy>=1.11
matplotlib>=3.8
